requirements.txt
fastapi[standard]
pydantic
torch
torchvision
pillow
requests